Rmbg-Xez/requirements.txt
Flask==2.3.3
rembg==2.0.31
Pillow==9.3.0
